requirements.txt
# Core dependencies
torch==2.2.0
torchaudio==2.2.0
transformers==4.37.2
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
pyyaml==6.0.1
webargs~=5.5.3
deepaas>=2.1.0

# For Hugging Face audio models (CLAP)
tokenizers==0.15.1
huggingface-hub==0.20.3
safetensors==0.4.2

# Optional but likely useful
tqdm==4.66.1
scikit-learn==1.4.0  # used in many pipelines
